SPORTS-22 Move tables constants file
