+++ chess/backend/src/constants/tables.ts
