+++ chess/backend/src/models/tournament.ts
@@ -1,5 +1,5 @@
 import BookShelf from '../config/bookshelf';
-import { TABLES } from './../utils/constants';
+import { TABLES } from './../constants/tables';
 
 class Tournament extends BookShelf.Model<Tournament> {
   get tableName() {
